+++ WorldEditArt/src/LegendsOfMCPE/WorldEditArt/Epsilon/Consts.php
@@ -66,4 +66,5 @@ class Consts{
 	const PERM_SELECT_DEFAULT = "worldeditart.builder.select.default";
 	const PERM_SELECT_DESEL = "worldeditart.builder.select.desel";
 	const PERM_SELECT_SET_CUBOID = "worldeditart.builder.select.set.cuboid";
+	const PERM_SELECT_SET_CYLINDER = "worldeditart.builder.select.set.cylinder";
 }