+++ WorldEditArt/src/LegendsOfMCPE/WorldEditArt/Epsilon/Selection/Wand/Impl/CylinderTopCenterWand.php
@@ -19,57 +19,42 @@ namespace LegendsOfMCPE\WorldEditArt\Epsilon\Selection\Wand\Impl;
 
 use LegendsOfMCPE\WorldEditArt\Epsilon\BuilderSession;
 use LegendsOfMCPE\WorldEditArt\Epsilon\Consts;
-use LegendsOfMCPE\WorldEditArt\Epsilon\IShape;
 use LegendsOfMCPE\WorldEditArt\Epsilon\Selection\Wand\AbstractFieldDefinitionWand;
+use LegendsOfMCPE\WorldEditArt\Epsilon\UserInterface\UserFormat;
 use pocketmine\level\Position;
-use pocketmine\utils\TextFormat;
 use sofe\libgeom\Shape;
-use sofe\libgeom\shapes\CuboidShape;
+use sofe\libgeom\shapes\CircularFrustumShape;
 
-class CuboidPosWand extends AbstractFieldDefinitionWand{
-	/** @var bool */
-	private $one;
-
-	public function __construct(bool $one){
-		$this->one = $one;
-	}
-
+class CylinderTopCenterWand extends AbstractFieldDefinitionWand{
 	public function getName() : string{
-		return $this->one ? "pos1" : "pos2";
+		return "cyltop";
 	}
 
 	public function getAliases() : array{
-		return $this->one ? ["1"] : ["2"];
+		return ["ct"];
 	}
 
 	public function getPermission() : string{
-		return Consts::PERM_SELECT_SET_CUBOID;
+		return Consts::PERM_SELECT_SET_CYLINDER;
 	}
 
 	protected function canModify(Shape $shape) : bool{
-		return $shape instanceof CuboidShape;
+		return $shape instanceof CircularFrustumShape;
 	}
 
 	protected function modify(BuilderSession $session, Shape $shape, Position $position){
-		assert($shape instanceof CuboidShape);
-		if($this->one){
-			$shape->setFrom($position);
+		assert($shape instanceof CircularFrustumShape);
+		$oldTop = $shape->getTop();
+		$shape->setTop($position);
+		if($oldTop !== null){
+			$session->msg("Moved top center from " . UserFormat::formatVector($oldTop) . " to " . UserFormat::formatVector($position), BuilderSession::MSG_CLASS_SUCCESS);
 		}else{
-			$shape->setTo($position);
+			$session->msg("Set top center to " . UserFormat::formatVector($position));
 		}
 	}
 
-	protected function createNew(Position $position) : IShape{
-		$shape = new class($position->getLevel()) extends CuboidShape implements IShape{
-			public function getBaseShape() : Shape{
-				return $this;
-			}
-		};
-		if($this->one){
-			$shape->setFrom($position);
-		}else{
-			$shape->setTo($position);
-		}
-		return $shape;
+	protected function canCreateNoisy(BuilderSession $session) : bool{
+		$session->msg("Please define the base of the cylinder/cone/circular frustum before defining the top", BuilderSession::MSG_CLASS_ERROR);
+		return false;
 	}
 }