+++ WorldEditArt/src/LegendsOfMCPE/WorldEditArt/Epsilon/UserInterface/Commands/WeaStatusCommand.php
@@ -18,7 +18,7 @@ declare(strict_types=1);
 namespace LegendsOfMCPE\WorldEditArt\Epsilon\UserInterface\Commands;
 
 use LegendsOfMCPE\WorldEditArt\Epsilon\Consts;
-use LegendsOfMCPE\WorldEditArt\Epsilon\UserInterface\PlayerBuilderSession;
+use LegendsOfMCPE\WorldEditArt\Epsilon\Session\PlayerBuilderSession;
 use LegendsOfMCPE\WorldEditArt\Epsilon\WorldEditArt;
 use pocketmine\command\CommandSender;
 use pocketmine\Player;